+++ ast/src/parser/identifier.rs
@@ -1,4 +1,8 @@
-use crate::lexer::Position;
+use crate::parser::lex::{
+    lexer::{Position, LocatedToken, Request},
+    token::Token,
+};
+use crate::parser::base::{State, Result, Error};
 use std::hash::{Hash, Hasher};
 
 #[derive(Debug, Clone)]
@@ -13,6 +17,29 @@ pub type Variable = Identifier;
 pub type FunctionName = Variable;
 // Identifier that starts with uppercase letter.
 pub type ConstructorName = Identifier;
+
+// TODO: These are more appropriate for lexer.
+pub fn identifier(state: &mut State) -> Result<Identifier> {
+    let LocatedToken { token: Token::Identifier(str), position } = state.request_token(Request::Identifier)? else { unreachable!() };
+    let identifier = Identifier::new(str, position);
+    Ok(identifier)
+}
+
+pub fn constructor_name(state: &mut State) -> Result<ConstructorName> {
+    let id = identifier(state)?;
+    if !id.first_char().is_ascii_uppercase() { return Err(Error::ExpectedTypeConstructor { received: id }) }
+    Ok(id)
+}
+
+pub fn variable(state: &mut State) -> Result<Variable> {
+    let id = identifier(state)?;
+    if !id.first_char().is_ascii_lowercase() { return Err(Error::ExpectedTypeConstructor { received: id }) }
+    Ok(id)
+}
+
+pub fn function_name(state: &mut State) -> Result<FunctionName> {
+    variable(state)
+}
 
 impl Identifier {
     pub fn new(str: String, position: Position) -> Self {
